实现了创建房间、删除房间的接口方法 实现了根据分店编号删除房间的接口方法 现在删除分店会同时删除其房间
不重要：
  实现了通过分店编号获取房间编号的服务
  可以跨微服务调用实现的第二个方法

+++ hotel-service/hotel-service-branch/src/main/java/com/abcd/branch/controller/BranchController.java
@@ -33,7 +33,7 @@ public class BranchController {
     private RoomFeignClient roomFeignClient;
 
     /**
-     * 创建分店，分店名称、地址、电话不能为空
+     * 创建分店
      * /api/branch/save
      */
     @Operation(summary = "创建分店")
@@ -50,12 +50,17 @@ public class BranchController {
     }
 
     /**
-     * 删除分店，需要指定分店编号
+     * 删除分店
      * /api/branch/delete/{branchId}
      */
     @Operation(summary = "删除分店")
     @DeleteMapping("/delete/{branchId}")
     public ResponseResult removeBranch(@PathVariable Integer branchId) throws Exception {
+
+        ResponseResult responseResult = roomFeignClient.removeRoomsByBranchId(branchId);
+
+        if(responseResult.getCode() != 1)
+            return ResponseResult.error("删除分店房间失败!");
 
         boolean result = branchService.removeById(branchId);
